The response to a DELETE request now includes a header.

+++ source/job/Resource_delete.cpp
@@ -2,11 +2,26 @@
 
 using job::Resource;
 
+static std::string	make_headers();
+
 http::Status
 Resource::_delete(route::Location const& loc) {
 	stdfs::path const&	pt = loc.to();
 
+	if (!std::filesystem::exists(pt))
+		return (http::Status::not_found);
 	if (!std::filesystem::remove(pt))
 		return (http::Status::internal_error);
+	_iss.clear();
+	_iss.str(make_headers());
 	return (http::Status::no_content);
+}
+
+static std::string
+make_headers() {
+	std::ostringstream	oss;
+
+	oss << "Connection: keep-alive\r\n";
+	oss << "\r\n";
+	return (oss.str());
 }
